requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
colorama==0.4.6
Django==4.2.2
gunicorn==21.2.0
packaging==23.1
Pillow==10.0.0
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.4.4
typing-extensions==4.6.3
tzdata==2023.3

